title: 60fps sub-frame interpolation for the 30fps cs — camera spline, rider XZ+Y, fog eye, fireglow (#149)
The ported title-cs consumers were pure functions of the integer cs frame,
which ticks once per two engine frames (sTickParity) — camera, rider position
and fog stepped at half rate ('everything jitters, not like 60fps').
New Zelda3D_TitleCsSubframe() (0.0 advance tick / 0.5 hold tick) feeds:
- Zelda3D_TitleCsCamera: float frame, fractional track eval (integer segment select)
- TitleRider::applyToActor: half-step XZ advance + per-engine-frame floor raycast Y
- fog-eye caller and fireglow cmab sampling at the fractional frame
Rider MOVE log now prints world pos + anim frame for per-engine-frame verification.

Also: tools/zelda3d_game.sh now defaults ZELDA3D_HEADLESS=1 (agent tooling must
never open a desktop window; ZELDA3D_HEADLESS=0 is the explicit headed opt-out).

+++ tools/soh3d_harness/main.cpp
@@ -166,7 +166,7 @@ int  Zelda3D_TitleCsEndFrame(void);
 // zelda3d_cutscene.cpp — ported byte-exact OP97 camera spline evaluator;
 // title-sync's eye-inversion (DeriveAzLockCsByEyeInversion) inverts the
 // oracle's live camera eye against this to derive its held cs frame.
-int  Zelda3D_TitleCsCamera(int frame, float eye[3], float at[3], float up[3],
+int  Zelda3D_TitleCsCamera(float frame, float eye[3], float at[3], float up[3],
                            float* fovDeg);
 
 extern uint8_t*     gSoh3dCaptureBuf;
